refactor: 위젯 역활별 분리 및 파일 경로 수정
